+++ src/Fable.Cli/Main.fs
@@ -459,14 +459,19 @@ type ProjectCracked
     member _.MapSourceFiles(f) =
         ProjectCracked(cliArgs, crackerResponse, Array.map f sourceFiles)
 
-    static member Init(cliArgs: CliArgs) =
+    static member Init(cliArgs: CliArgs, ?evaluateOnly: bool) =
+        let evaluateOnly = defaultArg evaluateOnly false
         Log.always $"Parsing {cliArgs.ProjectFileAsRelativePath}..."
 
         let result, ms =
             Performance.measure
             <| fun () ->
-                CrackerOptions(cliArgs)
+                CrackerOptions(cliArgs, evaluateOnly)
                 |> getFullProjectOpts (MSBuildCrackerResolver())
+        // TODO: Should we have a switch to support both resolvers?
+        // or should we go with the MSBuild one and remove the other one?
+        // CrackerOptions(cliArgs, evaluateOnly)
+        // |> getFullProjectOpts (BuildalyzerCrackerResolver())
 
         // We display "parsed" because "cracked" may not be understood by users
         Log.always
@@ -1247,7 +1252,10 @@ let private compilationCycle (state: State) (changes: ISet<string>) =
                     let oldProjCracked = projCracked
 
                     let newProjCracked =
-                        ProjectCracked.Init({ cliArgs with NoCache = true })
+                        ProjectCracked.Init(
+                            { cliArgs with NoCache = true },
+                            evaluateOnly = true
+                        )
 
                     // If only source files have changed, keep the project checker to speed up recompilation
                     let fableCompiler =