+++ tests/Js/Main/ArrayTests.fs
@@ -1199,4 +1199,10 @@ let tests =
         xs |> equal [|1; 2; 3; 0; 0; 0; 0|]
         Array.Resize(&xs, 0)
         xs |> equal [||]
+        xs <- null
+        Array.Resize(&xs, 3)
+        xs |> equal [|0; 0; 0|]
+        xs <- null
+        Array.Resize(&xs, 0)
+        xs |> equal [||]
   ]